requirements.txt
numpy
scipy
matplotlib
jaxopt
